reorganized the code and made the inner for loop in run a method of its own.

+++ include/Arrival.h
@@ -22,7 +22,7 @@ class PoissonArrival
         double                      A;
         std::random_device          rd;
         std::mt19937                gen;
-        std::uniform_real_distribution<int> d;
+        std::uniform_real_distribution<double> d;
 };
 
 #endif
